Add `contextIsolation: false` to node setup to enable `require`
As of Electron v12, `contextIsolation` is enabled by default: https://www.electronjs.org/docs/breaking-changes#default-changed-contextisolation-defaults-to-true. This prevents you from calling `require` in the render process. This quick fix ensures that the Realm quickstart can be built with newer version of Electron - it might make sense to refactor it to work with the default of `contextIsolation` enabled in future.