(DOCSP-27518): Reorg flutter realm database docs (#2509)
## Pull Request Info

Reorg flutter realm database docs.

### why?

the Flutter SDK TOC was getting really bloated, and is going to get more
so with merge of https://github.com/mongodb/docs-realm/pull/2507. did
this PR to consolidate the TOC and clearly group all Realm DB related
info together.

As part of this change:

- moved all Realm Database-related Flutter docs to Realm Database
section.
- split up the `realm-database.rst` include file so i could put
different parts of it in Flutter SDK
- fixed some existing redirect issues
- redirects!

> **Warning**
> 
> Dear reviewer, pls pay extra attention to the redirect changes.
> i have a tendancy of messing up redirect-related changes 😅 

### Jira

- https://jira.mongodb.org/browse/DOCSP-27518

### Staged Changes

- [Realm Database (Flutter) (whole section)
](https://docs-mongodbcom-staging.corp.mongodb.com/realm/docsworker-xlarge/DOCSP-27518/sdk/flutter/realm-database)

### Reminder Checklist

If your PR modifies the docs, you might need to also update some
corresponding
pages. Check if completed or N/A.

- [x] Create Jira ticket for corresponding docs-app-services update(s),
if any
- [x] Checked/updated Admin API
- [x] Checked/updated CLI reference

### Review Guidelines


[REVIEWING.md](https://github.com/mongodb/docs-realm/blob/master/REVIEWING.md)