(DOCSP-26730): C++: Add a React to Changes page (#2570)
## Pull Request Info

Note for reviewer: Git shenanigans resulted in #2568 being closed when
the feature branch it was working against got merged. This PR was
approved in #2568 but had a couple of pieces of feedback, which I
incorporated here.

### Jira

- https://jira.mongodb.org/browse/DOCSP-26730

### Staged Changes

- [React to
Changes](https://docs-mongodbcom-staging.corp.mongodb.com/realm/docsworker-xlarge/DOCSP-26730/sdk/cpp/react-to-changes/)

### Reminder Checklist

If your PR modifies the docs, you might need to also update some
corresponding
pages. Check if completed or N/A.

- [x] Create Jira ticket for corresponding docs-app-services update(s),
if any
- [x] Checked/updated Admin API
- [x] Checked/updated CLI reference

### Review Guidelines


[REVIEWING.md](https://github.com/mongodb/docs-realm/blob/master/REVIEWING.md)