(no Jira) Remove Flipper images (#2698)
## Pull Request Info

As part of the screenshot audit, I removed images on the React Native
page "Debugging with Flipper". These images are liable to be quickly
outdated.

Made a few minor text updates to not refer directly to images.

### Staged Changes

- [Debugging with
Flipper](https://docs-mongodbcom-staging.corp.mongodb.com/realm/docsworker-xlarge/remove-flipper-images/sdk/react-native/test-and-debug/debugging-with-flipper/)

### Reminder Checklist

If your PR modifies the docs, you might need to also update some
corresponding
pages. Check if completed or N/A.

- [x] Create Jira ticket for corresponding docs-app-services update(s),
if any
- [x] Checked/updated Admin API
- [x] Checked/updated CLI reference

### Review Guidelines


[REVIEWING.md](https://github.com/mongodb/docs-realm/blob/master/REVIEWING.md)

### Animal wearing a hat


![](https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcS_s2bO8RbjT6PgQmJfbNmbJs00KmZCTiDXMMX1vaF2AT-BazGyiuKVCfRwI_CYUerjvd8&usqp=CAU)