Update Expo Template References (#2715)
We are going TS only now for the templates. In that light, we have
created `@realm/expo-template` to be a typescript template and will not
be updating `expo-template-js` or `expo-template-ts` anymore.

## Pull Request Info

### Jira

- https://jira.mongodb.org/browse/DOCSP-NNNNN

### Staged Changes

-
[PAGE_NAME](https://docs-mongodbcom-staging.corp.mongodb.com/realm/docsworker-xlarge/BRANCH_NAME/)

### Reminder Checklist

If your PR modifies the docs, you might need to also update some
corresponding
pages. Check if completed or N/A.

- [ ] Create Jira ticket for corresponding docs-app-services update(s),
if any
- [ ] Checked/updated Admin API
- [ ] Checked/updated CLI reference

### Review Guidelines


[REVIEWING.md](https://github.com/mongodb/docs-realm/blob/master/REVIEWING.md)