(DOCSP-28375) Add info about opening a realm with no schema (#2753)
## Pull Request Info

Adds info and tests/examples about how to open a realm without passing
in schema as part of the config object. Also, refactored some
unnecessary includes for the Node.js page.

The Node.js test is failing in CI, but works locally. This test file
needs a bunch of work, and the Node.js CI is flaky. I can look into as
part of this PR if we think it's necessary, but I'd rather cut a ticket
to focus on fixing all the tests in that file.

### Jira

- https://jira.mongodb.org/browse/DOCSP-28375

### Staged Changes

- React Native: [Configure a
Realm](https://docs-mongodbcom-staging.corp.mongodb.com/realm/docsworker-xlarge/DOCSP-28375/sdk/react-native/realm-files/configure-a-realm/#access-a-realm-without-providing-a-schema)
- Node.js: [Open & Close a
Realm](https://docs-mongodbcom-staging.corp.mongodb.com/realm/docsworker-xlarge/DOCSP-28375/sdk/node/realm-files/open-and-close-a-realm/#open-a-realm-without-providing-a-schema)

### Review Guidelines


[REVIEWING.md](https://github.com/mongodb/docs-realm/blob/master/REVIEWING.md)

### Animal Wearing a Hat


![](https://4.bp.blogspot.com/_zFfOR5W31y4/SOx65YR0_JI/AAAAAAAAAmo/JcsHhDBa7lY/w1200-h630-p-k-no-nu/krakatoaAP_450x300.jpg)

---------

Co-authored-by: Dachary <dc@dacharycarey.com>