(DOCSP-30205) Add information about `isCleared` (#2830)
## Pull Request Info

I debated whether to include this property or not. We don't currently
document it. However, it is listed in the API reference docs. I decided
it makes sense to include it even though it's deprecated in the latest
release.

### Jira

- https://jira.mongodb.org/browse/DOCSP-30205

### Staged Changes

- [React to
Changes](https://docs-mongodbcom-staging.corp.mongodb.com/realm/docsworker-xlarge/DOCSP-30205/sdk/flutter/realm-database/react-to-changes/#register-a-query-change-listener)

### Review Guidelines


[REVIEWING.md](https://github.com/mongodb/docs-realm/blob/master/REVIEWING.md)

### Animal Wearing a Hat

<img src="https://i.redd.it/te7a93pgry461.jpg" width=300>