(DOCSP-29619): Map a Class to a Different Persisted Name (#2965)
## Pull Request Info

The Jira ticket also requests adding this in Java and React Native,
which I'm not doing in this ticket. We're generally not updating the
Java docs except for bugs, so that's a won't do. With the v12 release it
probably makes sense to do the React Native stuff in v12, so I'll split
that to a separate ticket for later. This functionality is not available
in Swift, and already has examples in Flutter - I've just cross-posted
them for this ticket.

### Jira

- https://jira.mongodb.org/browse/DOCSP-29619

### Staged Changes

- [Node.js - Define an Object
Model](https://docs-mongodbcom-staging.corp.mongodb.com/realm/docsworker-xlarge/DOCSP-29619/sdk/node/model-data/define-a-realm-object-model/#map-a-property-or-class-to-a-different-name)
- [Flutter - Define an Object
Model](https://docs-mongodbcom-staging.corp.mongodb.com/realm/docsworker-xlarge/DOCSP-29619/sdk/flutter/realm-database/model-data/define-realm-object-schema/#map-a-model-or-class-to-a-different-name)
- [Flutter - Property
Annotations](https://docs-mongodbcom-staging.corp.mongodb.com/realm/docsworker-xlarge/DOCSP-29619/sdk/flutter/realm-database/model-data/property-annotations/#map-a-property-or-class-to-a-different-name)

### Reminder Checklist

If your PR modifies the docs, you might need to also update some
corresponding
pages. Check if completed or N/A.

- [x] Create Jira ticket for corresponding docs-app-services update(s),
if any
- [x] Checked/updated Admin API
- [x] Checked/updated CLI reference

### Review Guidelines


[REVIEWING.md](https://github.com/mongodb/docs-realm/blob/master/REVIEWING.md)