DOCSP-32246: Remove instances of `kotlin-sdk` role (#2969)
## Pull Request Info

https://github.com/mongodb/snooty-parser/pull/506 will update
`kotlin-sdk` role to point to the base library. This replaces any
current instances of the `kotlin-sdk` role with source constants, so
links don't break after that PR merges.

[DOCSP-32519](https://jira.mongodb.org/browse/DOCSP-32519) will replace
all source constants with the newly defined kotlin roles after the merge
🎉

### Jira

- https://jira.mongodb.org/browse/DOCSP-32246

### Staged Changes

-
https://docs-mongodbcom-staging.corp.mongodb.com/realm/docsworker-xlarge/docsp-32246-kotlin-sdk-role/

### Reminder Checklist

If your PR modifies the docs, you might need to also update some
corresponding
pages. Check if completed or N/A.

- [x] Create Jira ticket for corresponding docs-app-services update(s),
if any
- [x] Checked/updated Admin API
- [x] Checked/updated CLI reference

### Review Guidelines


[REVIEWING.md](https://github.com/mongodb/docs-realm/blob/master/REVIEWING.md)