(DOCSP-32576) Document new subscribe API for React Native (#3010)
## Pull Request Info

Note: the bulk of the lines changed in the PR are due to
`package-lock.json` changes - so not something that needs to be
reviewed.

This PR adds examples and descriptions for the new Subscribe API.
Updating the older subscription examples is out of scope. This PR
already turned into more work than expected.

I ran into some less-than-ideal behavior with how Realm and React Native
interact and needed to refactor the tests and file structure with that
behavior in mind.

Specifically:
- Object model classes should be defined outside of components where a
realm is opened. In this case, that means anywhere `RealmProvider` is
used. Due to how React Native's hot reload works, the realm stays open,
but the class reference changes. The result is this error:
  > Constructor was not registered in the schema for this Realm

Re-starting the simulated app fixes the problem, but breaking hot reload
is a terrible dev experience.
- Providers should also be defined in their own wrapper file. If you
don't, you could run into a similar issue introduced by re-rendering and
get this error:
	> Access to invalidated Results objects

Check out this issue for details:
https://github.com/realm/realm-js/issues/6066
- I re-structured the test app's navigation a bit to allow for different
pages within the Subscriptions area.
- I also moved the `__tests__` directory from the project root into
`src`, which is a more common pattern.

One of the new tests consistently fails in CI, but works locally just
fine. It takes more than 10 seconds to run on CI, which exceeds the
limit I've set for Jest. I'll try to figure out why it's taking so long.

### Jira

- https://jira.mongodb.org/browse/DOCSP-32576

### Staged Changes

- [Manage Sync
Subscriptions](https://docs-mongodbcom-staging.corp.mongodb.com/realm/docsworker-xlarge/DOCSP-32576/sdk/react-native/sync-data/flexible-sync/)

### Review Guidelines


[REVIEWING.md](https://github.com/mongodb/docs-realm/blob/master/REVIEWING.md)

### Animal Wearing a Hat

<img
src="https://i.chzbgr.com/full/9197265920/h0EE55220/bearded-dragon-with-top-hat-reptile"
width=400>