(DOCSP-32633) Node.js update test suite object models (#3033)
## Pull Request Info

Loooots of changes to generated snippets in this PR. Most of them are
very minor changes to snippets that aren't object models. This comes
from standardizing automated formatting.

I also cleaned up some references to Bluehawk snippets that no longer
exist.

### Jira

- https://jira.mongodb.org/browse/DOCSP-32633
- https://jira.mongodb.org/browse/DOCSP-32626

### Staged Changes

There are a lot of changes to snippets across the Node.js SDK docs. The
vast majority of them are minor. I didn't look at all of them in
staging, but if you have the desire feel free. 😅

-
[Node.js](https://preview-mongodbkrollinsmdb.gatsbyjs.io/realm/DOCSP-32633-node/sdk/node/)

### Review Guidelines


[REVIEWING.md](https://github.com/mongodb/docs-realm/blob/master/REVIEWING.md)

### Animal Wearing a Hat

<img src="https://i.chzbgr.com/full/9563697408/h4E1A70CA/hat" width=400>