(DOCSP-30804): Flutter: Break out CRUD section, List and Set asResults() (#3121)
## Pull Request Info

Jira ticket: https://jira.mongodb.org/browse/DOCSP-30804

-
[CRUD/Read](https://preview-mongodbdacharyc.gatsbyjs.io/realm/DOCSP-30804/sdk/flutter/crud/read/#convert-lists-or-sets-to-results):
New "Convert Lists or Sets to Results" section with tested, Bluehawked
code example demonstrating the functionality.

As part of this PR, I also broke out the Read & Write Data page into a
new CRUD section. I put it at the root of the SDK because a future
iteration will remove the "Realm Database" docs from the "Realm
Database" section of the SDK. Putting it at the root of the SDK means we
won't have to make redirects for this section when that time comes -
we'll only have to update the ToC.

I copied language that has already been approved in the C++ and Kotlin
SDKs to flesh out CRUD details. The only changes I made to that language
was to update the SDK docs links referenced, and to omit Realm naming
when possible in favor of the new naming/positioning.

### Reminder Checklist

Before merging your PR, make sure to check a few things.

- [x] Did you tag pages appropriately?
  - genre
  - meta.keywords
  - meta.description
- [x] Describe your PR's changes in the Release Notes section
- [x] Create a Jira ticket for related docs-app-services work, if any

### Release Notes

- **Flutter SDK**
- Realm/Realm Database/Read & Write Data: Break out the single page to a
new CRUD section.
- Realm/Realm Database/CRUD/Read: New "Convert Lists or Sets to Results"
section with tested, Bluehawked code example demonstrating the
functionality.

### Review Guidelines


[REVIEWING.md](https://github.com/mongodb/docs-realm/blob/master/REVIEWING.md)