Add refs to easily link to telemetry pages from Edge Server docs (#3234)
## Pull Request Info

In a PR in App Services, I'm adding a breakdown of different types of
telemetry reported when using Edge Server. I'd like to link out to Atlas
Device SDK telemetry documentation so people who are using SDK clients
with Edge Server can easily see the telemetry info for an SDK.

Not adding a release note here because it's such a trivial internal
thing that it doesn't seem worth reporting.