Rename the Web package.json to stop Snyk PRs (#3374)
## Pull Request Info

After investigation to the Snyk integration, it seems we are unable to
disable it for a specific directory or repository.

While we are evaluating long-term plans for the Web SDK test suite,
getting unnecessary Snyk update PRs is disruptive and creates
maintenance burden.

This PR renames the `package.json` file, which is one of the files that
Snyk scans for, which should hopefully stop the Snyk dependency update
PRs. The updated README provides instructions about the changes required
to run the test suite locally and reinstate it in the future.