(DOCSP-48512): Add redirect from introduction to deprecation (#3390)
Jira ticket: https://jira.mongodb.org/browse/DOCSP-48512

The former `Introduction` page had some outdated messaging. It has been
removed from the repo, but was still available when browsed to via URL.
Adding a redirect to make sure the outdated messaging is not available
and it points to the correct content.